BUG: Replace itk::Math::Round template argument `double` with `int64_t`
According to https://github.com/InsightSoftwareConsortium/ITK/blob/66ab5b14c476d057ed46a4daba05b710681160e6/Modules/Core/Common/include/itkMath.h#L127 the `TReturn` template argument of `Math::Round` must be an integer type. In practice, it will internally call `Detail::RoundHalfIntegerToEven_64` when `TReturn` is a 64-bit `double`. Which returns an `int64_t` anyway.

+++ Common/CostFunctions/itkAdvancedImageToImageMetric.hxx
@@ -550,7 +550,7 @@ AdvancedImageToImageMetric<TFixedImage, TMovingImage>::EvaluateMovingImageValueA
         MovingImageIndexType index;
         for (unsigned int j = 0; j < MovingImageDimension; ++j)
         {
-          index[j] = static_cast<long>(Math::Round<double>(cindex[j]));
+          index[j] = static_cast<long>(Math::Round<int64_t>(cindex[j]));
         }
         (*gradient) = this->m_GradientImage->GetPixel(index);
       }
